+++ R/CovariateSettings.R
@@ -158,7 +158,7 @@ createAgeCovariateSettings <- function(ageKnots = 5,
 #' Create an object specifying whether and how seasonality should be included in the model.
 #' Seasonality can be included by splitting patient time into calendar months. During a month, the
 #' relative risk attributed to season is assumed to be constant, and the risk from month to month is
-#' modeled using a cyclic cubic spline.
+#' modeled using a cyclic quadratic spline.
 #'
 #' @param seasonKnots                  If a single number is provided this is assumed to indicate the
 #'                                     number of knots to use for the spline, and the knots are